requirements-new.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.6
aioredis==2.0.1
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.21.3
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
Authlib==1.6.0
bidict==0.23.1
black==23.12.1
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.3
dataclasses-json==0.6.7
Deprecated==1.2.18
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
fastapi==0.115.12
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.5.1
google-adk==1.1.1
google-ai-generativelanguage==0.4.0
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.95.1
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.23.3
google-cloud-speech==2.32.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.1
google-crc32c==1.7.1
google-genai==1.18.0
google-generativeai==0.3.2
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.20.3
greenlet==3.2.2
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.62.3
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.3
identify==2.6.12
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.1.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.1.5
langchain-community==0.0.20
langchain-core==0.1.23
langsmith==0.0.87
marshmallow==3.26.1
mcp==1.9.2
motor==3.3.2
multidict==6.4.4
mypy==1.8.0
mypy_extensions==1.1.0
nodeenv==1.9.1
numpy==1.26.4
openai==1.10.0
opentelemetry-api==1.33.1
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
pre-commit==3.6.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==4.25.8
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pymongo==4.13.0
pyparsing==3.2.3
pytest==7.4.4
pytest-asyncio==0.23.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-engineio==4.12.1
python-multipart==0.0.20
python-socketio==5.11.0
PyYAML==6.0.2
redis==5.0.1
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
shapely==2.1.1
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sse-starlette==2.3.6
starlette==0.46.2
structlog==24.1.0
tenacity==8.2.3
threadpoolctl==3.6.0
tokenizers==0.15.2
tqdm==4.67.1
transformers==4.37.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
virtualenv==20.31.2
watchfiles==1.0.5
websockets==15.0.1
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.0
zipp==3.22.0
